[LOD-178] feat: 보관함 전송에 따른 재화 감소 refac: 재화변경 금액 Integer -> Long 상품 테이블 count->stock_quantity API 호출에 따른 변경 사항

+++ src/main/java/net/lodgames/shop/product/vo/ProductOptionVo.java
@@ -13,6 +13,6 @@ public class ProductOptionVo {
     private Long productId;          // 상품 아이디
     private String name;             // 이름
     private ProductOptionType type;  // 상품 옵션 타입 DIAMOND(0), COIN(1), ITEM(2)
-    private Integer quantity;        // 갯수
+    private Long quantity;           // 갯수
     private LocalDateTime createdAt; // 생성시각
 }
